Added cast to parameter and modified GetBytesByName type
I've modified some variable type and function in order to be able to compile and execute the code on my machine. This version is working perfectly. Maybe other people hadthe same issue ? 

For instance "GetBytesByName" was a bool but returned nothing (compilation error).
"OffsetNamesTable" variable was a DWORD* leading to a compilation error during the operation  base + OffsetNamesTableOffset (compilation error)